pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/donorpipe"]

[project]
name = "donorpipe"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "console-menu>=0.8.0",
    "fastapi[standard]>=0.135.1",
    "python-dateutil>=2.9.0.post0",
]

[project.scripts]
donorpipe = "donorpipe.cli:main"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyright>=1.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
